# server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Torrent streaming server.",
        "title": "Swagger Torrserver API",
        "contact": {},
        "license": {
            "name": "GPL 3.0"
        },
        "version": "{version.Version}"
    },
    "basePath": "/",
    "paths": {
...
        "/msx/imdb/{id}": {
            "get": {
                "description": "Get MSX IMDB informations.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "MSX"
                ],
                "summary": "Get MSX IMDB informations",
                "parameters": [
                    {
                        "type": "string",
                        "description": "IMDB ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "JSON MSX IMDB informations"
                    }
                }
            }
        },
        "/msx/{pth}": {
            "get": {
                "description": "Multi usage endpoint.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "MSX"
                ],
                "summary": "Multi usage endpoint",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Route MSX pages",
                        "name": "link",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Data returned according to path"
                    }
                }
            }
        },
...
    },
...
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        }
    },
...
}